flask/pyproject.toml
[project]
name = "flask-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.3",
    "pika==1.3.2",
    "python-dotenv==1.0.1",
]
